Moved css to a main index file. Issue #6
