+++ src/components/Game.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import '../styles/index.css';
-import '../styles/button-controls.css';
+import '../index.css';
 import * as Utils from '../utils/gameUtils.js';
 import * as Constants from '../constants.js';
 import { getScoreBoardThunk } from '../store/thunkCreators/gameThunkCreators.js';
